+++ changelog.md
@@ -1,3 +1,4 @@
 - Added directional goat horn sounds
 - Made music disc hearing range configurable
 - Made music disc range in line with vanilla music discs
+- Added config options to disallow uploading mp3s or wavs